requirements.txt
# 1. Core Dependencies
setuptools==68.0.0
wheel>=0.38.0
pip>=23.0.1
numpy==1.21.6
colorama==0.4.6
pytz==2023.3

# 2. Machine Learning Core
torch==2.4.1+cu128
torchvision==0.15.2+cu128
torchaudio==2.0.2+cu128
flash-attn==2.5.6

# 3. Database and Environment
redis==5.0.0
python-dotenv==0.21.0
pandas==2.0.0
aiofiles==24.1.0

# 4. Audio Processing
portaudio==19.6.0
pyaudio==0.2.13
pynput==1.7.6
pydub==0.25.1
soundfile==0.12.1
wave==0.0.2
SpeechRecognition==3.10.0
sounddevice==0.4.5
librosa==0.10.0
pygame==2.6.1

# 5. Machine Learning & AI
transformers==4.31.0
sentence-transformers==3.4.1
timm==0.9.2
wandb==0.15.5
protobuf==3.19.6
openai==1.63.0
TTS==0.22.0

# 6. Web & API
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
aiohttp==3.8.5
fastapi==0.115.8 
starlette==0.45.3 
uvicorn==0.34.0

# 7. Windows Specific
pywin32==306; platform_system == "Windows"

# 8. Visualization
matplotlib==3.5.3

# 9. Special Installations
whisper @ git+https://github.com/openai/whisper.git

# 10. Build Dependencies
packaging
ninja